refactor(rust): group the TLS args, and nest HttpTlsConfig into HttpConfig
HttpTlsConfigArgs carries the seven TLS/mTLS options (cert_pem,
key_pem, ca_cert, cert_p12, cert_p12_password, allow_unsafe_connection,
override_target_name). Unlike Tcp/Http2/Proxy, these share no common
prefix, so grouping is a plain flatten with no with_prefix!: every
name already matches today's flat one exactly.

The three certificate-reading helpers and the override-target
resolution (which needs the already-resolved endpoint, so resolve()
takes it as a parameter) move alongside, in the same file as their own
Args type. HttpConfig nests tls: HttpTlsConfig, completing the same
symmetry already applied to tcp/http2/proxy: every thematic unit now
groups its Args and resolved types together, and HttpConfig itself no
longer has any flat field that belongs to one of them.

Caught during validation: cargo clippy --no-default-features never
compiles #[cfg(test)] code, so a Secret import gated behind the serde
feature (correct for its one non-test use) silently broke
cargo test --no-default-features, where two tests reference Secret
unconditionally. Fixed by importing it directly inside the test
module, where the feature does not apply. cargo test
--workspace --no-default-features --no-run is now part of this
crate's validation, alongside the existing clippy/doc checks.
