+++ packages/rust/armonik-transport/src/env.rs
@@ -145,7 +145,7 @@ mod tests {
         for spelling in ["1", "true", "yes", "enable", "allow", "authorize"] {
             let config = resolve(spelling).expect(spelling);
             assert!(
-                config.allow_unsafe_connection,
+                config.tls.allow_unsafe_connection,
                 "`{spelling}` should read as true"
             );
         }
@@ -153,7 +153,7 @@ mod tests {
         for spelling in ["0", "false", "no", "disable", "disallow", "forbid", ""] {
             let config = resolve(spelling).expect(spelling);
             assert!(
-                !config.allow_unsafe_connection,
+                !config.tls.allow_unsafe_connection,
                 "`{spelling}` should read as false"
             );
         }
@@ -189,7 +189,7 @@ mod tests {
         )
         .expect("reading the variable itself must not fail");
 
-        assert_eq!(args.cert_pem, "no/such/cert.pem");
+        assert_eq!(args.tls.cert_pem, "no/such/cert.pem");
     }
 
     #[test]
@@ -198,7 +198,7 @@ mod tests {
         let args = HttpConfigArgs::from_env("ARMONIK_TEST_NOCERT__")
             .expect("an unset variable names no path");
 
-        assert_eq!(args.cert_pem, "");
+        assert_eq!(args.tls.cert_pem, "");
     }
 
     #[test]
@@ -292,7 +292,7 @@ mod tests {
         )
         .expect("the escape hatch this reader documents must work");
 
-        assert_eq!(args.override_target_name, "[::1]");
+        assert_eq!(args.tls.override_target_name, "[::1]");
     }
 
     #[test]
@@ -306,13 +306,13 @@ mod tests {
         })
         .expect("an empty variable must not fail the read");
 
-        assert_eq!(args.cert_pem, "");
+        assert_eq!(args.tls.cert_pem, "");
         let config = HttpConfig::from_config_args(HttpConfigArgs {
             endpoint: String::from("http://localhost:5001"),
             ..args
         })
         .expect("an empty cert_pem must not be treated as half an identity");
-        assert!(config.identity.is_none());
+        assert!(config.tls.identity.is_none());
     }
 
     #[test]
