+++ packages/rust/armonik-transport/src/tcp_config.rs
@@ -36,7 +36,7 @@ pub struct TcpConfigArgs {
     #[cfg_attr(feature = "serde", serde(deserialize_with = "text"))]
     pub keepalive_retries: String,
     /// Enable Nagle's algorithm (disable TCP_NODELAY), empty for false. `TcpNagleAlgorithm`. See
-    /// [`crate::HttpConfigArgs::allow_unsafe_connection`] for the accepted spellings.
+    /// [`crate::HttpTlsConfigArgs::allow_unsafe_connection`] for the accepted spellings.
     #[cfg_attr(feature = "serde", serde(deserialize_with = "text"))]
     pub nagle_algorithm: String,
 }
